move the event directive code from plasmapy_sphinx.extras to plasmapy_sphinx.directives.event; delete any references to extras
