--- docs/api_static/plasmapy_sphinx.extras.rst
@@ -1,8 +1,0 @@
-:orphan:
-
-`plasmapy_sphinx.extras`
-========================
-
-.. currentmodule:: plasmapy_sphinx.extras
-
-.. automodapi:: plasmapy_sphinx.extras
